feat: added new view of forms
