+++ app/Repositories/Forms/Form/EloquentFormRepository.php
@@ -7,15 +7,31 @@ use App\Repositories\Forms\Form\FormRepository;
 
 class EloquentFormRepository implements FormRepository
 {
-    public function getAllForm($request) {
-        return Forms::orderBy('date', 'desc')->get();
+    public function getAllForm($request)
+    {
+        $query = Forms::query();
+
+        if ($request->filled('search')) {
+            $search = $request->input('search');
+
+            $query->where(function ($q) use ($search) {
+                $q->where('title', 'like', '%' . $search . '%')
+                    ->orWhere('date', 'like', '%' . $search . '%')
+                    ->orWhere('status', 'like', '%' . $search . '%');
+            });
+        }
+
+        return $query->orderBy('created_at', 'desc')->paginate(5);
     }
 
-    public function getFormById($id) {
+
+    public function getFormById($id)
+    {
         return Forms::find($id);
     }
 
-    public function getActualForm() {
+    public function getActualForm()
+    {
         return Forms::where('status', '1')->first();
     }
